@@ -13,7 +13,7 @@
     "build:lib:shared": "ng build ngx-shared",
     "build:lib:shared:watch": "ng build ngx-shared --watch",
     "build:analyze": "npm run build:lib:shared && ng build --named-chunks --stats-json && webpack-bundle-analyzer dist/luv-coffee-fe/stats.json",
-    "build:start": "npm run build:local && cd dist/luv-coffee-fe && spa-server-gzip -p 4200",
+    "build:start": "npm run build:local && cd dist/luv-coffee-fe/browser && spa-server-gzip -p 4200",
     "lint": "ng lint",
     "lint:fix": "ng lint --fix",
     "lint:app": "ng lint luv-coffee-fe",
@@ -28,30 +28,30 @@
     "prepush": "npm run test"
   },
   "dependencies": {
-    "@angular/animations": "^17.3.10",
-    "@angular/cdk": "^17.3.10",
-    "@angular/common": "^17.3.10",
-    "@angular/compiler": "^17.3.10",
-    "@angular/core": "^17.3.10",
-    "@angular/forms": "^17.3.10",
-    "@angular/material": "^17.3.10",
-    "@angular/platform-browser": "^17.3.10",
-    "@angular/platform-browser-dynamic": "^17.3.10",
-    "@angular/router": "^17.3.10",
+    "@angular/animations": "^18.0.0",
+    "@angular/cdk": "^18.0.0",
+    "@angular/common": "^18.0.0",
+    "@angular/compiler": "^18.0.0",
+    "@angular/core": "^18.0.0",
+    "@angular/forms": "^18.0.0",
+    "@angular/material": "^18.0.0",
+    "@angular/platform-browser": "^18.0.0",
+    "@angular/platform-browser-dynamic": "^18.0.0",
+    "@angular/router": "^18.0.0",
     "@angularnestpostgre/packages": "^0.0.3",
     "rxjs": "~7.4.0",
     "tslib": "^2.3.0",
     "zone.js": "~0.14.6"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.8",
+    "@angular-devkit/build-angular": "^18.0.1",
     "@angular-eslint/builder": "^17.5.1",
     "@angular-eslint/eslint-plugin": "^17.5.1",
     "@angular-eslint/eslint-plugin-template": "^17.5.1",
     "@angular-eslint/schematics": "^17.5.1",
     "@angular-eslint/template-parser": "^17.5.1",
-    "@angular/cli": "^17.3.8",
-    "@angular/compiler-cli": "^17.3.10",
+    "@angular/cli": "^18.0.1",
+    "@angular/compiler-cli": "^18.0.0",
     "@types/jasmine": "~3.10.0",
     "@types/node": "^12.11.1",
     "@typescript-eslint/eslint-plugin": "^7.10.0",
@@ -63,7 +63,7 @@
     "karma-coverage": "~2.0.3",
     "karma-jasmine": "~4.0.0",
     "karma-jasmine-html-reporter": "~1.7.0",
-    "ng-packagr": "^17.3.0",
+    "ng-packagr": "^18.0.0",
     "typescript": "~5.4.5"
   }
 }
